+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [UNRELEASED] - 0000-00-00
 ### Fixed
+- Validate certificate in `tls set` command
+- Sorted keys in `datacol env`
 - Showing request/limit for memory, cpu for a process 
 - Bump default GKE cluster version to `1.8.8-gke.0`
 - Bump GCP image to `debian-8-jessie-v20170523` version
@@ -14,6 +16,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `datacol destroy <name>` don't need `STACK` env var
 - CLI to respect `--app` and `--stack` flag
 ### Added
+- TLS certificates for GCP based stacks
 - Index for `status` and `created_at` for datastore Build object
 - Colored output for `datacol env`
 - [CLI] `datacol env unset` supports multiple parameters